backend/requirements.txt
Flask
Flask-CORS
python-dotenv
PyJWT
pymongo
bcrypt
authlib
pyyaml
scikit-learn
spacy
google-genai
sentence-transformers
torch
pandas
openpyxl


